fix(contrato): corregir tipo de dato del porcentaje a entero en la validación del contrato

+++ app/config/static/js/view/contrato.js
@@ -76,7 +76,7 @@ const valorContrato = document.getElementById("valorContrato");
             // console.log("Valor del contrato:", e.target.value);
             
             const valorContrato = parseFloat(document.getElementById("valorContrato").value);
-            const porcentaje = parseFloat(document.getElementById("porcentaje").value);
+            const porcentaje = parseInt(document.getElementById("porcentaje").value);
         
             if(porcentaje < 0 || isNaN(porcentaje)) {
                 alert("Porcentaje de interés no puede ser negativo o inválido.");
